+++ src/main/java/net/setrion/koratio/world/entity/demonic/DemonicSkeleton.java
@@ -47,7 +47,7 @@ public class DemonicSkeleton extends Skeleton {
     }
 
     public static AttributeSupplier.Builder createAttributes() {
-        return Monster.createMonsterAttributes().add(Attributes.FOLLOW_RANGE, 35.0D).add(Attributes.MAX_HEALTH, 20.0D).add(Attributes.MOVEMENT_SPEED, (double)0.27F).add(Attributes.ATTACK_DAMAGE, 3.0D);
+        return Monster.createMonsterAttributes().add(Attributes.FOLLOW_RANGE, 35.0D).add(Attributes.MAX_HEALTH, 20.0D).add(Attributes.MOVEMENT_SPEED, 0.27F).add(Attributes.ATTACK_DAMAGE, 3.0D);
     }
 
     public boolean canBreakDoors() {
